Auto Weather Driving
I have added the two driving scripts that we care about during
inital testing to a subdirectory.  At some point, I will need to
make a new script with side-by-side displays, but that is a future
task.

manual_control is just the provided driving script with Carla.

data_gather is based on manual_control.  It only uses the necessary
sensors, and it automatically begins driving in first-person mode
with constantly and randomly switching weather.  The next task will
be recording pictures (and then possibly moving the weather switching
before pygame.flip or the carla simulator computataion, depending on
how the pictures turn out when I take them back-to-back).