Implement Image Conversion Scripts
Since the output dataset from Carla doesn't match the specifications
used by RainCityScapes, for either depth (inversion) or semantic
segmentation (different labeling standard), I created a pair of
Python scripts to automatically convert a folder of data into the
alternative format.  The conversion scripts use the PIL library and
store the data in a subdirectory called converted_images/ of the
directory that the script is run from (I think -- it could just be
an ignored subdirectory of ConversionScripts, because I didn't
test that).  Be aware that it isn't the fastest, but it runs in a
watchable amount of time.