Create Carla Segmentation Demo
I edited the automatic_control example to drive the ego vehicle
around and display the output of Detectron.  I also edited Detectron
some to run on my computer (and run alongside Carla, because the
two together both running on my GPU runs CUDA out of memory, which
is ... unfortunate).

I did modify the segmentation sensor listener to stop listening
during processing and call listen() again before leaving the
listener, but that was to keep the data closer to current, instead
of having a backlog, which was sort-of how it looked before.