Create Script to Change Scene
So far, all of the collected data had been in the default town.
However, I was concerned that this would result in overfitting, so
I looked up how to load a different (built-in) map into the
simulator.  It has been implemented as change_town.py, which takes
one digit as a command line argument.